Trigger on published release, show status in body
Trigger on publish or prerelease of a release. The script attempts to
checkout and build a qt/qt5 repository version, with the same tag as the
one in the release, i.e. must be tagged with something like: "v5.14.1".

Use github-script-s to add a 'pending' badge at startup and to replace
it with 'passed' or 'failed' badges at completion.